pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "shahnoor77",email = "shahnoorkhan9955@gmail.com"}
]
readme = "README.md"

requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "mysql-connector-python (>=9.5.0,<10.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "prometheus-client (>=0.23.1,<0.24.0)",
    "tenacity (>=9.1.2,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "ruff (>=0.14.7,<0.15.0)",
    "mypy (>=1.19.0,<2.0.0)"
]
